check whether first string argument of variadic template logging has format specifiers
first, std::vsnprintf is used to see whether the (format) string with
all arguments results in a string with a different length than the
(format) string itself. if that's the case, we assume that formatting
worked out and forward again to std::vsnprintf to call the logging
backend. Otherwise a regex for format specifier is used to match the
first string argument - if it matches, we again proceed as before,
otherwise, logging is done via a stream.

closes #3